refactoring, adding helper methods for occupancy processor
